🚚 [RUMF-1046] rename xhrProxy module to xhrObservable
